Update project files for .NET platform and C#
The previous versions did not work, but after tweaking the code, the GUI
window is finally appearing with an interactable button. This program
outputs a single text expressing my everlasting love to my girlfriend,
but it will serve as the backbone for more functionalities. The other
buttons will connect to a chatGPT model and provide output.
To be continued.